Fix rendering of long lines in narrow browser
These CSS changes fix the rendering of long lines when the browser is horizontally constrained (i.e. a phone screen width or a narrow window on a desktop OS).

For example, before this change, if you attempted to browse [Return a File Stored in Google Cloud Storage via a Play Application](https://andrew.selvia.com/journal/software/developer/return-a-file-stored-in-google-cloud-storage-via-a-play-application.xhtml) on a phone, the code elements would define the width of the entire page since they were the longest elements. In practice, this meant that the paragraphs were rendered mostly off-screen to the right. Reading articles that fell under this category was burdensome. With these changes, the content should lay out within the boundaries of the viewport and the code can simply be scrolled horizontally.

Additionally, I tested how pages with deeply nested lists appeared. If that happens, the same principle applies: the body text will render within the boundaries of the viewport. Unlike the code case, however, the deeply-nested list items will continue to be rendered to the right. After some thought, I think this problem is unavoidable. Medium doesn't even let its articles have nested list items, so at least I'm doing better than that.